requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
requests==2.28.1
openai==0.27.0
Werkzeug==2.2.2